app/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-jose
passlib
tenacity
slowapi
